requirements.txt
pandas==2.1.4
numpy==1.26.3
PyYAML==6.0.1
